requirements.txt
beautifulsoup4==4.5.1
dateparser==0.5.0
eventlet==0.19.0
feedparser==5.2.1
futures==3.0.5
greenlet==0.4.10
html5lib==0.999999999
jdatetime==1.8.1
pymongo==3.3.1
python-dateutil==2.5.3
pytz==2016.7
regex==2016.10.22
requests==2.11.1
ruamel.ordereddict==0.4.9
ruamel.yaml==0.12.15
schedule==0.3.2
six==1.10.0
toolz==0.8.0
typing==3.5.2.2
tzlocal==1.3
umalqurra==0.2
urllib3==1.18.1
webencodings==0.5
xmltodict==0.10.2
